deagg class reorganization
