+++ src/org/opensha2/calc/GroundMotions.java
@@ -31,8 +31,6 @@ final class GroundMotions {
    *
    * TODO It would be nice to have an immutable variant of a double[] backed
    * list, but would require copying values on build().
-   *
-   * TODO refactor to μLists σLists
    */
 
   final InputList inputs;
